news-crawler-agent/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
feedparser>=6.0.0
trafilatura>=1.4.0
python-slugify>=8.0.0
google-cloud-storage>=2.9.0
httpx>=0.23.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
lxml-html-clean>=0.1.0